tests/fixtures/autogen-group/requirements.txt
pyautogen>=0.2.0
openai
